requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0rc0
google-api-python-client==2.108.0
google-auth==2.40.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
tqdm==4.67.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
